Stripped-down signal loop for timing purposes
Implements a simple counter-based signal loop on the PRU which doesn't actually
use the data copied in by the host, rather sends a fixed number of ones or
zeros.

For this I've found that a timing of T0H = 350ns and T1H = 750ns works well, as
specified in the WS2812 datasheet, except that a period of 2500ns appears to be
required (contrary to the datasheet).